feat: refactor and arrange code
